requirements.txt
langchain
langchain-anthropic
langchain-openai
langchain-community
langchain-core
langchain-voyageai
langchain-pinecone
langchainhub
jupyter
pandas
python-dotenv
google-search-results
numexpr
pinecone-client
pinecone
voyageai
anthropic
datasets